databaseCon/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
annotated-types=0.6.0=py312h06a4308_0
anyio=4.6.2=py312h06a4308_0
azure-core=1.29.5=py312h06a4308_0
azure-identity=1.17.1=pyhd8ed1ab_0
brotli-python=1.0.9=py312h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.9.24=h06a4308_0
certifi=2024.8.30=py312h06a4308_0
cffi=1.17.1=py312h1fdaa30_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py312h06a4308_0
cryptography=43.0.0=py312hdda0065_0
expat=2.6.3=h6a678d5_0
fastapi=0.112.2=py312h06a4308_0
greenlet=3.0.1=py312h6a678d5_0
h11=0.14.0=py312h06a4308_0
idna=3.7=py312h06a4308_0
ld_impl_linux-64=2.40=h12ee557_0
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
msal=1.25.0=py312h06a4308_0
msal_extensions=0.3.0=pyh9f0ad1d_0
ncurses=6.4=h6a678d5_0
openssl=3.0.15=h5eee18b_0
pip=24.2=py312h06a4308_0
portalocker=1.4.0=py_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=2.8.2=py312h06a4308_0
pydantic-core=2.20.1=py312hb02cf49_0
pyjwt=2.8.0=py312h06a4308_0
pyodbc=5.1.0=py312h6a678d5_0
pysocks=1.7.1=py312h06a4308_0
python=3.12.7=h5148396_0
python-dotenv=1.0.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=py312h06a4308_0
setuptools=75.1.0=py312h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=py312h06a4308_0
sqlalchemy=2.0.34=py312h00e1ef3_0
sqlite=3.45.3=h5eee18b_0
starlette=0.38.2=py312h06a4308_0
tk=8.6.14=h39e8969_0
typing-extensions=4.11.0=py312h06a4308_0
typing_extensions=4.11.0=py312h06a4308_0
tzdata=2024b=h04d1e81_0
unixodbc=2.3.11=h5eee18b_0
urllib3=2.2.3=py312h06a4308_0
uvicorn=0.20.0=py312h06a4308_0
wheel=0.44.0=py312h06a4308_0
xz=5.4.6=h5eee18b_1
zlib=1.2.13=h5eee18b_1
